Consolidate ts editing functions to tests.tsutil.
