+++ tests/test_stats.py
@@ -23,11 +23,11 @@ from __future__ import print_function
 from __future__ import division
 
 import unittest
-import random
 
 import numpy as np
 
 import msprime
+import tests.tsutil as tsutil
 
 
 def get_r2_matrix(ts):
@@ -59,7 +59,6 @@ def get_r2_matrix(ts):
     return A
 
 
-@unittest.skip("Replace calls  to copy()")
 class TestLdCalculator(unittest.TestCase):
     """
     Tests for the LdCalculator class.
@@ -138,14 +137,14 @@ class TestLdCalculator(unittest.TestCase):
 
     def test_single_tree_simulated_mutations(self):
         ts = msprime.simulate(20, mutation_rate=10, random_seed=15)
-        sites = sorted(random.sample(list(ts.sites()), self.num_test_sites))
-        ts = ts.copy(sites)
+        ts = tsutil.subsample_sites(ts, self.num_test_sites)
         self.verify_matrix(ts)
         self.verify_max_distance(ts)
 
     @unittest.skip("recurrent mutations")
     def test_single_tree_regular_mutations(self):
         ts = msprime.simulate(self.num_test_sites, length=self.num_test_sites)
+        # TODO change this to use branch_mutations from tsutil.
         sites = [
             msprime.Site(
                 position=j, index=j, ancestral_state="0",
@@ -160,23 +159,21 @@ class TestLdCalculator(unittest.TestCase):
             self.num_test_sites, recombination_rate=1,
             length=self.num_test_sites)
         self.assertGreater(ts.get_num_trees(), 10)
-        sites = [
-            msprime.Site(
-                position=j, index=j, ancestral_state="0",
-                mutations=[msprime.Mutation(
-                    site=j, node=j, derived_state="1", parent=msprime.NULL_MUTATION)])
-            for j in range(self.num_test_sites)]
-        ts = ts.copy(sites)
+        t = ts.dump_tables()
+        t.sites.reset()
+        t.mutations.reset()
+        for j in range(self.num_test_sites):
+            site_id = len(t.sites)
+            t.sites.add_row(position=j, ancestral_state="0")
+            t.mutations.add_row(site=site_id, derived_state="1", node=j)
+        ts = msprime.load_tables(**t.asdict())
         self.verify_matrix(ts)
         self.verify_max_distance(ts)
 
     def test_tree_sequence_simulated_mutations(self):
         ts = msprime.simulate(20, mutation_rate=10, recombination_rate=10)
         self.assertGreater(ts.get_num_trees(), 10)
-        sites = random.sample(list(ts.sites()), self.num_test_sites)
-        ts = ts.copy(sorted(sites, key=lambda x: x.position))
-        # for site in ts.sites():
-        #     print(site)
+        ts = tsutil.subsample_sites(ts, self.num_test_sites)
         self.verify_matrix(ts)
         self.verify_max_distance(ts)
         self.verify_max_mutations(ts)
